requirements.txt
python-dotenv
langchain-core
langchain-openai
langchain-experimental
langchain-community
langchain-google-genai
langchain-cohere
transformers 
huggingface_hub
langchain-huggingface
langchain-anthropic
openai
pandas
langchain-anthropic
vllm